+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - BPay handling to NSW GovPay provider
+- `formsAppService.getFormSubmissionMetaList()`
 
 ## [10.1.2] - 2026-05-07